app/requirements.txt
psycopg2-binary==2.9.1
SQLAlchemy==1.4.26
speedtest-cli
requests